Move: PhotoListCard, VoucherListCard를 @component/list/ 경로에 배치하고, 공통적으로 사용되는 PhotoCard 컴포넌트는 @component/photocard/ 경로에 배치함
