+++ client/src/component/list/photo/photoListCardSlice.ts
