+++ client/src/component/photocard/SkeletonPhotoCard.tsx
